Add a basic summary of submissions for each survey
