+++ app/routes/surveys/components/surveys.css
@@ -218,3 +218,7 @@
 .answers > li {
   margin: 10px 0 20px;
 }
+
+.submissionNav {
+  margin: 10px 0;
+}
