+++ app/routes/surveys/index.js
@@ -22,7 +22,19 @@ export default {
     },
     {
       path: ':surveyId/submissions',
-      ...resolveAsyncRoute(() => import('./SubmissionsRoute'))
+      ...resolveAsyncRoute(() => import('./SubmissionsRoute')),
+      childRoutes: [
+        {
+          path: 'summary', // admin/groups/123/settings
+          ...resolveAsyncRoute(() => import('./components/SubmissionSummary'))
+        },
+        {
+          path: 'individual', // admin/groups/123/members
+          ...resolveAsyncRoute(() =>
+            import('./components/SubmissionIndividual')
+          )
+        }
+      ]
     }
   ]
 };
